+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- filterArray function.
 ## [6.2.0] - 2022-02-24
 ### Added
 - objectToObjectsArray function.